Oidc refactor candidate
